Move wire protocol definitions into their own file
